requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
numpy==2.1.2
openai==1.52.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dispatch==0.2.2
python-dotenv==1.0.1
python-osc==1.9.0
scipy==1.14.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
websocket-client==1.8.0
